[tracing] TS: Add addNestedTags to replace jaegerClient.addTags
