+++ components/gitpod-protocol/package.json
@@ -13,6 +13,7 @@
         "@types/analytics-node": "^3.1.5",
         "@types/chai-subset": "^1.3.3",
         "@types/cookie": "^0.4.1",
+        "@types/jaeger-client": "^3.18.3",
         "@types/js-yaml": "^3.10.1",
         "@types/mocha": "^5.2.7",
         "@types/node": "^16.11.6",
